Move custom handler to `src/logging`.
