+++ src/gui/main_window.py
@@ -16,7 +16,13 @@ import pyqtgraph as pg
 from collections import OrderedDict
 from copy import deepcopy
 
-from src.gui.dialog_box import ConnectWindow, OptionsWindow, FilterWindow, AboutWindow, ExportWindow
+from src.gui.dialog_box import (
+    ConnectWindow,
+    OptionsWindow,
+    FilterWindow,
+    AboutWindow,
+    ExportWindow,
+)
 from src.gui.widgets import CustomLabel, CustomLineEdit, CustomDockWidget
 
 from src.gui.equity_chart import EquityChart
